Changed price field from float to double in Ride and its two concrete classes.
Added ThemePark class which has responsibility for managing operations such as checking height for customers against restrictions for rides. ThemePark has lists of Rides and Customers.

+++ src/main/java/Dodgem.java
@@ -1,7 +1,7 @@
 public class Dodgem extends Ride{
 
 
-    public Dodgem(String type, float price, int minimumAge, int minimumHeightInCentimetres) {
+    public Dodgem(String type, double price, int minimumAge, int minimumHeightInCentimetres) {
         super(type, price, minimumAge, minimumHeightInCentimetres);
     }
